@@ -20,7 +20,7 @@
     "@fortawesome/free-brands-svg-icons": "^5.4.2",
     "@fortawesome/free-solid-svg-icons": "^5.1.0",
     "@fortawesome/vue-fontawesome": "^0.1.0",
-    "@nuxtjs/axios": "5.1.1",
+    "@nuxtjs/axios": "^5.3.6",
     "@riophae/vue-treeselect": "0.0.37",
     "cookie": "^0.3.1",
     "element-ui": "^2.4.11",
@@ -32,13 +32,13 @@
     "lodash.uniq": "^4.5.0",
     "lodash.uniqby": "^4.7.0",
     "moment": "^2.22.2",
-    "nuxt": "^1.4.2",
-    "nuxt-fontawesome": "^0.3.0",
+    "nuxt": "^2.3.4",
+    "nuxt-fontawesome": "^0.4.0",
     "string": "^3.3.3",
     "string-mask": "^0.3.0",
     "vue-avatar": "^2.1.4",
     "vue-clickaway": "^2.2.2",
-    "vue-i18n": "7.6.0",
+    "vue-i18n": "^8.7.0",
     "vue-scrollto": "^2.11.0",
     "vue-the-mask": "^0.11.1",
     "vue2-transitions": "^0.2.3",
@@ -46,18 +46,18 @@
     "vuex-persistedstate": "^2.5.4"
   },
   "devDependencies": {
-    "babel-eslint": "8.2.1",
+    "babel-eslint": "^10.0.1",
     "babel-plugin-component": "^1.1.1",
-    "eslint": "4.15.0",
-    "eslint-friendly-formatter": "3.0.0",
-    "eslint-loader": "1.7.1",
-    "eslint-plugin-vue": "4.0.0",
+    "eslint": "^5.11.1",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-loader": "^2.1.1",
+    "eslint-plugin-vue": "^5.1.0",
     "node-sass": "^4.9.4",
-    "postcss-loader": "2.1.1",
+    "postcss-loader": "^3.0.0",
     "sass-loader": "^7.1.0",
     "vue-avatar-component": "^1.2.4",
-    "vue-loader": "^13.7.2",
-    "vue-svg-loader": "^0.5.0"
+    "vue-loader": "^15.4.2",
+    "vue-svg-loader": "^0.11.0"
   },
   "engines": {
     "node": "9.3.0",
